refactor: isolate query param restoration
